move t type to the separate file
